refactor(react-await): separate useAwait, AwaitClient
