+++ packages/react-await/src/index.ts
@@ -1,4 +1,7 @@
-export { Await, useAwait, awaitClient } from './Await'
+export { Await } from './Await'
+export { useAwait } from './useAwait'
 export { awaitOptions } from './awaitOptions'
+export { awaitClient } from './AwaitClient'
 
-export type { AwaitProps, AwaitOptions } from './Await'
+export type { AwaitProps } from './Await'
+export type { AwaitOptions } from './types'
